requirements.txt
paho-mqtt
pykka
python-telegram-bot
psycopg2
requests